backend/requirements.txt
asgiref==3.7.2
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.1.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rest-condition==1.0.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
gunicorn==20.1.0
psycopg2-binary==2.9.3